Split https scanner into utils and https
