+++ platform/overlays/knative/images/dmarc_reports/requirements.txt
@@ -2,3 +2,4 @@ Flask
 waitress
 PyJWT
 requests>=2.18.4
+gql
